Move to helper and use new helper
