+++ open_alchemy/helpers/foreign_key.py
@@ -164,11 +164,9 @@ def get_modify_schema(
     assert type_ != relationship.Type.MANY_TO_MANY
 
     if type_ == relationship.Type.ONE_TO_MANY:
-        items_schema = peek.items(schema=property_schema, schemas=schemas)
-        assert items_schema is not None
-        ref = peek.ref(schema=items_schema, schemas=schemas)
-        assert ref is not None
-        _, ref_schema = ref_helper.get_ref(ref=ref, schemas=schemas)
+        _, ref_schema = relationship.get_ref_schema_many_to_x(
+            property_schema=property_schema, schemas=schemas
+        )
         return ref_schema
 
     return parent_schema
@@ -200,11 +198,9 @@ def get_modify_name(
     assert type_ != relationship.Type.MANY_TO_MANY
 
     if type_ == relationship.Type.ONE_TO_MANY:
-        items_schema = peek.items(schema=property_schema, schemas=schemas)
-        assert items_schema is not None
-        ref = peek.ref(schema=items_schema, schemas=schemas)
-        assert ref is not None
-        ref_name, _ = ref_helper.get_ref(ref=ref, schemas=schemas)
+        ref_name, _ = relationship.get_ref_schema_many_to_x(
+            property_schema=property_schema, schemas=schemas
+        )
         return ref_name
 
     return parent_name
