+++ open_alchemy/schemas/helpers/property_/__init__.py
@@ -1,6 +1,5 @@
 """Helpers for properties."""
 
-from . import relationship
 from . import simple
 from . import type_
